backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
numpy
pandas
xgboost
joblib
transformers
torch
sentence-transformers
faiss-cpu
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-community
openai
anthropic
google-generativeai
PyMuPDF
pdfplumber
python-multipart
python-dotenv
httpx
aiofiles